DateUtils: introduce timeZoneContext
This is a context manager which temporary allow to change the time zone.

It is not thread safe and intended to be used in tests only, to replace
setTimeZoneToUTC and setTimezone methods by something unified which
makes the transition to DateTime 4 easier, because changing timezone
was done by patching private attributes and the DateTime 4 use different
attributes.