km: use context free document lookup in "document renderer"
This gadget looks up a document by reference, by using
getDocumentValueList on the context of the web section.
getDocumentValueList also applies the criterions from the context, but
because web page references are already unique, adding more criterions
can just cause the document to be not found and the gadget to display
the "No document" error.