erp5_core_test: restricted: Test pandas resampler
In some projects we use pandas resampler inside 'portal_callables'
(so with restricted Python). But we miss tests which check
whether pandas resamplers are allowed inside the restricted Python.
Due to this lack of tests we won't notice whether the usage of pandas
resamplers still work with a newer pandas version. With this patch
we can recognize if an internal change of pandas makes the resampler unusuable
in restricted Python.

/reviewed-on https://lab.nexedi.com/nexedi/erp5/merge_requests/1738
/reviewed-by @jerome @klaus