erp5_core_test: restricted: Test pandas resampler
We use in projects pandas resampler inside 'portal_callables'
(so with restricted Python). But we miss tests which check
whether pandas resamplers are allowed inside the restricted Python.
Due to this lack we couldn't notice if they still worked with a newer
pandas version. With this patch we can recognize if an internal
change of pandas makes the resampler unusuable in restricted Python.