CMFActivity,syncml: stop using a dedicated log file
$INSTANCE_HOME/log only exists in runUnitTest instances, this code had
no effect in slapos instances.

It is a future problem because it depends on the Signals module (part
of ZServer package) and we are removing this dependency in the python 3
port.

For the SynchronizationTool, it was even more problematic because it
can cause the component not to loed, with an error like:

    ImportError: erp5.component.tool.SynchronizationTool: cannot load Component SynchronizationTool (signal only works in main thread)