ui_test: ignore errors with cell/getId for testAccessUnauthorizedRelationValue
This listbox can be used to show documents that user is not allowed to
access, so this `cell/getId` can raise. In Zope2 it was raising
Unauthorized and this Unauthorized was catched, on Zope4 it is raising
NotFound, which is no longer catch since 850e436e27 (
Products.ERP5Form.Form: Let NotFound exceptions propagate during
rendering, 2022-11-25), so this NotFound was raised as a site error,
which caused an error in the test.