Products.CMFActivity.ActivityTool: Stop relying on portal_catalog's owner
The intent here is to have an all-allowed user which (mainly ?) can
executed grouped activities, as there is no good user to switch to when
executing these (activities in a group may have been spawned by multiple
users). This is typically what we want, as for example catalog indexations
must produce the same output independently from the security context which
spawned them.
There is no reason to give any special meaning to the owner of a particular
tool: instead, just pick our de-facto standard all-powerful user.