accounting_l10n_fr: fix ValidDate computation for FEC
According to https://www.legifrance.gouv.fr/codes/article_lc/LEGIARTI000027804775/
ValidDate is the date when the transaction was validated ( "16. La date
de validation de l'écriture comptable"), this implementation was simply
returning the modification date, which was not always correct, for
example after causality state was recaculated.

This keeps the `getModificationDate` fallback for cases where FEC is
used with some non validated transactions, in order to keep compatibility