Localizer,tests: fix translationContext that was working only in test
Test setup monkey-patched some function which no longer existed on Zope4
this function was used in translationContext, which was working fine
in the test because the function was here thanks to the patch, but
failed with ImportError when used for real.

This fixes the import to use the real getRequest and drop the patch to
prevent similar problems.