Localizer,tests: fix translationContext that was working only in test
Test setup monkey-patched some function which we removed when merging
Zope4 support. This function was used in translationContext, it was
working fine in the tests, because the function was here thanks to the
monkey-patch, but it failed with ImportError when used for real.

This fixes the import to use the real getRequest and drop the patch to
prevent similar problems.