erp5_web_renderjs_uid: Improve checkbox rendering
Displaying nothing when a checkbox is not checked makes the rendering
confusing: the field's title may appear as an affirmation, which can be
the exact opposite to the current system's state.
So display an empty ballot box: U+2610 (BALLOT BOX).
Then, for consistency, when the checkbox is checked, switch to a character
which also contains a ballot box.
Thierry decided in favour of U+1F5F9 (BALLOT BOX WITH BOLD CHECK) rather
than U+2611 (BALLOT BOX WITH CHECK), so this is what this change is doing.

For more on this:
  https://en.wikipedia.org/wiki/Check_mark#Unicode