erp5_web_renderjs_uid: Improve checkbox rendering
Displaying nothing when a checkbox is not checked makes the rendering
confusing: the field's title may appear as an affirmation, which can be
the exact opposite to the current system's state.
So display an empty ballot box: U+2610 (BALLOT BOX).
Then, for consistency, when the checkbox is checked, switch to a character
which also contains a ballot box: BALLOT BOX WITH CHECK.

For more on this:
  https://en.wikipedia.org/wiki/Check_mark#Unicode