oauth_google_login: apply a 10 seconds timeout when connecting to google server
In case of network problems, the http requests made to google during
login may take too long and cause global unavailability of the ERP5
instance.

Today we saw in long request logs:

    2023-05-29 07:10:35,662 - Thread 140596157511424: Started on 1685336511.6; Running for 524.1 secs; Same.

oauth2client does not expose an API to set a timeout, but it allows
passing a custom http instance where we can set a timeout.