openid_connect,google_login: increase cache duration to 10 days
Because these two plugins implement the refresh of the authentication
token (unlike the facebook plugin), we can use a longer cache here.
With the previous value of 1 day, when not using the site for more than
1 day, token would expire from cache and user have to authenticate again.
Because the scenario of "access is revoked from authorization server"
is handled by refreshing the access token, we can use a longer duration
here, so that user do not have to login again after not using for two
days.