ERP5TypeTestCase: review usage of erp5_site_global_id
This needs to be set on portal_memcached, because it's no longer acquired
since 4889d523f9 (Define property's default value at class level., 2012-12-31)

Use a hash to get a shorter value, because memcached keys are limited
in size and a base64 of connection string was more than 70 characters.

Fix the data type for python3, this is expected to be str, not bytes.

XXX I feel setting on portal_memcached like this is wrong, I don't
understand 4889d523f9 and it seems to me that we should revert that
change now and let erp5_site_global_id be acquired from the ERP5 Site.
portal_caches also uses erp5_site_global_id and it acquires it.