property_sheets: don't set a default state_variable
Every time they add a new workflow, developer must carefully think and
decide what should be the state variable in this context. If we provide
a default value, developer might skip this step and just use the default.

This also had a bad side effect of making interaction_workfow.state_var
be simulation_state, it uses to be "state" and some code was using this
as a way to detect if a workflow is a standard workflow or an
interaction workflow.

Now that we no longer have a default value (DCWorkflow also had a
default value of "state"), adjust accessor generation to not generate
state variable accessors when no state variable is defined.