accounting: show mirror account in grouping dialog
A typical problem which prevents grouping is that the line is used in an
internal invoice and while all the conditions for grouping seems OK for
one side, they are not OK for the other side, because the accounts are
different from the mirror side. By showing the mirror account it should
be a bit easier for users to figure out what is wrong.