core: don't show list of items in initial business template install
During initial install in most cases we just want to install everything,
showing the listbox to choose install options is not necessary. The
problem is that this listbox is really big when installing multiple
business templates.

This introduces a "detailed install" checkbox, that is unchecked by
default, to make it possible to show the list to choose from anyway.