officejs_support_request_ui_test: check service worker first
This might fix flaky "testLogin". One theory is that a service worker
request to fill the cache happens while the selenium browser logs out,
which cause browser to keeps the authentication cookie.

This test waits for service worker to have filled the cache. We also
change the name (00) to run it first, so that subsequent tests are
executed with the cache already pre-filled.