Zelenium: add an option to ignore snapshot tests
Because they depend on the browser version, the window size, the
available fonts and maybe other things, snapshot tests basically only
work when running on testnode and fail when we run the tests from our
browsers locally. With this option, the difference of snapshot are not
considered a failure.

Also adjust snapshots, because this adds a checkbox in the upper frame,
the viewport of the test frame is now a little bit taller.