dms: version up pdfjs 3.11.174
using pdfjs-3.11.174-legacy-dist.zip

Note: viewer.js is modified to do nothing in webViewerLoad (one line
with return "ERP5 patch: disable here" is added at the top of this
function) because we do the initialization in the gadget.
I think something similar was also done with the previous version.

The tests were extended to cover:

 - opening an empty or broken PDF: this should not crash
 - clicking save: this should not crash and the updated document should
   render properly. The save action replaces the ERP5 document content
   with the content from the editor, so if user for example added
   annotations or draw on the PDF, the modified version is saved.
   We don't use this feature and it's not so easy to test, so for now
   the test only covers the "it should not crash and it should not
   break the document".
 - entering wrong password for encrypted PDFs should not crash