monaco_editor: version up monaco-editor 0.44.0, integrate ruff
This integrates https://github.com/astral-sh/ruff compiled as web
assembly. For now it is enabled only for the gadget editor, so it works
for components.

The known problems are that it expects python3 syntax, so some python2
syntaxs will be reported as error, which is less and less a problem
this days. It is also enabled for ERP5 Python Scripts, but without the
necessary knowledge that a python script has arguments (that in this
version are reported as undefined names) and is actually a function body
so it complains when `return` is used.

This build was made with https://lab.nexedi.com/jerome/monaco-editor-erp5/commit/f914df70