ContributionTool: use getPropertyDictFromFilename to detect existing document
When a document is contributed via ContributionTool.newContent, there
is a check that no document already exist for the same reference, but
this was done using getMatchedFilenamePatternDict, which is not suitable
here because it just extract `reference` from the filename, but
getPropertyDictFromFilename, which is used in discoverMetadata [1] uses
a different approach and supports the reference to be captured by a
regexp capturing `node_reference` and `local_reference`.

By using the same approach to check the existing document, we prevent
an issue that the "You are not allowed to update the existing document"
was not raised when using a regex using `node_reference` and
`local_reference`.

Before the fix,
test_mergeRevision_with_node_reference_local_reference_filename_regular_expression
was failing and it passing after the fix. test_mergeRevision pass in
both cases, it was added because this scenario was apparently not tested.

1: https://lab.nexedi.com/nexedi/erp5/blob/4e48b8b6/product/ERP5/bootstrap/erp5_core/MixinTemplateItem/portal_components/mixin.erp5.DiscoverableMixin.py#L154