accounting: fix wrong mirror node displayed in grouping dialog
When account number method is set to account reference and the
transaction line has a node acquired from the invoice source or
destination organisation or person and when this node has a reference,
then this was displayed as an account reference.

We set a reference on Client 1 organisation to reproduce the problem.