dms: fix hidden buttons on PDF preview
Instead of disabling in javascript, which cause the button to be visible
for a few milliseconds, disable directly in the HTML file.

Do not hide "document properties", because it can be useful and it does
not hurt and "edit images" because we don't hide the buttons to edit
annotations either, so this is consistent.