core: drop text/comma-separated-values mime type
we use to have two entries for csv:

  - text/comma-separated-values
     - mimes: ["text/comma-separated-values"]
     - extensions: ["csv"]
     - globs: []
  - CSV document
     - mimes: ["text/csv", "text/x-comma-separated-values", "text/x-csv"]
     - extensions: []
     - globs: ["*.csv"]

but text/comma-separated-values does not really exist, rfc4180 recommends
text/csv.

The problem with this configuration is that when ERP5 picks a mime type
for csv extension, it uses text/comma-separated-values, as this one has
extensions set.

Change the configuration to delete "text/comma-separated-values" and keep
everything in "CSV document":

  - CSV document
     - mimes: ["text/csv", "text/x-comma-separated-values", "text/x-csv", "text/comma-separated-values"]
     - extensions: ["csv"]
     - globs: ["*.csv"]