ods_style: render all fields with style_macros/macro/cell_value
Instead of duplicating the code for the case where a field value is
rendered from a group (left, right etc) of the form and when the field
value is rendered from a listbox editable field, reuse the macro also
for fields from the form groups.

Because fields from the form groups have top and bottom border, we
introduce a new style prefix for these fields: "with_border_".

Implementation note: We don't declare this new field in
Base_getODSSupportedStylePrefixList, because we need to define the cell
properties each time, so defining this dynamically is not really
interesting.